core/api/paginations.py
from rest_framework.pagination import PageNumberPagination


class StandardPagination(PageNumberPagination):
    page_size = 200
    page_size_query_param = 'page_size'


class AllPagination(PageNumberPagination):
    page_size = 1000
