requirements.txt
Django==3.2.12
requests==2.27.1
django-cors-headers==3.13.0
django-extensions==3.1.5
djangorestframework==3.13.1
django-filter==21.1
dj-rest-auth==2.2.4
django-allauth==0.50.0
drf-extensions==0.7.1
drf-spectacular==0.26.1
django-anymail==8.6
django-waffle==3.0.0
jsonfield==3.1.0
gunicorn==20.1.0
redis==4.6.0
celery==5.3.1
numpy==1.24.4
scipy>=1.5.0
Pillow
django-jazzmin==2.6.0
python-dotenv
